+++ packages/form/src/components/CheckboxField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { act, screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { mockFormErrors, renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { CheckboxField } from '../..'
+import { mockFormErrors, renderWithForm } from '../../../__tests__/helpers'
 
 describe('checkboxField', () => {
   it('should render correctly', () => {
