+++ packages/form/src/components/Form/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import { screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('form', () => {
   it('renders correctly with node children', () => {
