+++ packages/form/src/components/KeyValueField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it } from 'vitest'
 import { KeyValueField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('keyValueField', () => {
   it('should render with default props', async () => {
