+++ packages/form/src/components/NumberInputField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { act, screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { mockFormErrors, renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { NumberInputField, Submit } from '../..'
+import { mockFormErrors, renderWithForm } from '../../../__tests__/helpers'
 
 describe('numberInputField', () => {
   it('should render correctly', () => {
