+++ packages/form/src/components/PhoneInputField/__tests__/index.test.tsx
@@ -1,9 +1,9 @@
 import { screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { mockFormErrors, renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { PhoneInputField } from '..'
 import { Submit } from '../..'
+import { mockFormErrors, renderWithForm } from '../../../__tests__/helpers'
 
 describe('form - PhoneInputField', () => {
   it('should render correctly', () => {
