+++ packages/form/src/components/RadioGroupField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { RadioGroupField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('radioField', () => {
   it('should render correctly checked', async () => {
