+++ packages/form/src/components/SelectableCardField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { act, screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { SelectableCardField } from '../..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('selectableCardField', () => {
   it('should render correctly', () => {
