+++ packages/form/src/components/SelectableCardGroupField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, vi, it } from 'vitest'
 import { SelectableCardGroupField } from '../..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('selectableCardField', () => {
   it('should render correctly', () => {
