+++ packages/form/src/components/TimeInputField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it, vi } from 'vitest'
 import { TimeInputField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('textInputField', () => {
   it('should render correctly', () => {
