+++ packages/form/src/components/ToggleField/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it } from 'vitest'
 import { ToggleField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('toggleField', () => {
   it('should render correctly', () => {
