+++ packages/form/src/components/ToggleGroupField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it, vi } from 'vitest'
 import { ToggleGroupField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('groupField', () => {
   it('should render correctly checked', async () => {
