+++ packages/form/src/components/VerificationCodeField/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it } from 'vitest'
 import { VerificationCodeField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 
 describe('verificationCodeField', () => {
   it('should render correctly', () => {
