+++ packages/form/src/compositions/OfferListField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it, vi } from 'vitest'
 import { OfferListField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 import { columns, data } from './resources'
 
 describe('offerListField', () => {
