+++ packages/form/src/compositions/PlansField/__tests__/index.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { renderWithForm } from '@utils/test'
 import { describe, expect, it, vi } from 'vitest'
 import { PlansField } from '..'
+import { renderWithForm } from '../../../__tests__/helpers'
 import { domain, fees, gb, pipeline, ssl } from './features'
 import { planAdvanced, planStarter } from './plans'
