+++ packages/form/src/compositions/RichTextInputField/__tests__/index.test.tsx
@@ -1,9 +1,10 @@
 import { renderHook, screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { mockFormErrors, renderWithForm, renderWithTheme } from '@utils/test'
+import { renderWithTheme } from '@utils/test'
 import { useForm } from 'react-hook-form'
 import { beforeAll, describe, expect, it, vi } from 'vitest'
 import { RichTextInputField } from '..'
+import { mockFormErrors, renderWithForm } from '../../../__tests__/helpers'
 import { Submit } from '../../../components'
 import { Form } from '../../../components/Form'
