+++ utils/test/src/index.ts
@@ -1,12 +1,3 @@
-export {
-  ComponentWrapper,
-  defaultError,
-  mockFormErrors,
-  mockMatchMedia,
-  renderWithForm,
-  renderWithTheme,
-  setup,
-  shouldMatchSnapshot,
-} from './vitest'
+export { ComponentWrapper, mockMatchMedia, renderWithTheme, resetUserEvent, setup, shouldMatchSnapshot } from './vitest'
 
 export * from './axe'
