+++ utils/test/src/vitest/index.ts
@@ -1,11 +1,4 @@
-export {
-  ComponentWrapper,
-  defaultError,
-  mockFormErrors,
-  renderWithForm,
-  renderWithTheme,
-  shouldMatchSnapshot,
-} from './helpers'
+export { ComponentWrapper, renderWithTheme, shouldMatchSnapshot } from './helpers'
 export { mockMatchMedia } from './mockMatchMedia'
 export { resetUserEvent } from './resetUserEvent'
 export { setup } from './setup'
